move parameters to route
