+++ router/lib/wabi/route.rb
@@ -1,11 +1,7 @@
 # frozen_string_literal: true
 
-require_relative 'params_from_path'
-
 module Wabi
   class Route
-    include ParamsFromPath
-
     attr_reader :http_verb, :path, :response
 
     def initialize(http_verb, path, response)
@@ -24,6 +20,24 @@ module Wabi
       else
         path == request_path
       end
+    end
+
+    def params_for_request(request)
+      params_from_path(request.path_info)
+        .merge(request.params)
+        .transform_keys(&:to_sym) # TODO: make deep symbolize
+    end
+
+    private
+
+    def params_from_path(request_path)
+      request_path
+        .match(named_capture_regex)
+        .named_captures
+    end
+
+    def named_capture_regex
+      Regexp.new(path.gsub(/:(?<param>\w+)/, '(?<\k<param>>\w+)'))
     end
   end
 end
